requirements.txt
openai>=1.0.0
mem0ai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
jinja2>=3.1.0
numpy>=1.24.0

